+++ features/support/env.rb
@@ -17,6 +17,8 @@ require 'site_prism'
 require 'capybara/poltergeist'
 require 'axe-capybara'
 require 'axe-cucumber-steps'
+require 'rake'
+require 'database_cleaner-active_record'
 
 # Require files we've created to help with the setup
 require_relative '../support/pages'
@@ -49,33 +51,11 @@ World(Pages)
 #
 ActionController::Base.allow_rescue = false
 
-# Remove/comment out the lines below if your app doesn't have a database.
-# For some databases (like MongoDB and CouchDB) you may need to use :truncation instead.
-begin
-  DatabaseCleaner.strategy = :transaction
-rescue NameError
-  raise 'You need to add database_cleaner to your Gemfile (in the :test group) if you wish to use it.'
-end
+# For running the rake task to relaod the test data
+Rails.application.load_tasks
 
-# You may also want to configure DatabaseCleaner to use different strategies for certain features and scenarios.
-# See the DatabaseCleaner documentation for details. Example:
-#
-#   Before('@no-txn,@selenium,@culerity,@celerity,@javascript') do
-#     # { except: [:widgets] } may not do what you expect here
-#     # as Cucumber::Rails::Database.javascript_strategy overrides
-#     # this setting.
-#     DatabaseCleaner.strategy = :truncation
-#   end
-#
-#   Before('not @no-txn', 'not @selenium', 'not @culerity', 'not @celerity', 'not @javascript') do
-#     DatabaseCleaner.strategy = :transaction
-#   end
-#
-
-# Possible values are :truncation and :transaction
-# The :transaction strategy is faster, but might give you threading problems.
-# See https://github.com/cucumber/cucumber-rails/blob/master/features/choose_javascript_database_strategy.feature
-Cucumber::Rails::Database.javascript_strategy = :transaction
+# Because we need to restore some data for JavaScript tests we need to invoke the Database strategy ourselves
+Cucumber::Rails::Database.autorun_database_cleaner = false
 
 # Capybara settings can go here
 Capybara.ignore_hidden_elements = false