+++ spec/helpers/header_navigation_links_helper_spec.rb
@@ -65,7 +65,7 @@ RSpec.describe HeaderNavigationLinksHelper do
         it 'returns the sign out link with the right options' do
           expect(result).to eq(
             [
-              { text: 'Sign out', href: '/facilities-management/sign-out', attributes: { method: :delete } }
+              { text: 'Sign out', href: '/facilities-management/sign-out', method: :delete }
             ]
           )
         end
@@ -78,7 +78,7 @@ RSpec.describe HeaderNavigationLinksHelper do
         it 'returns the sign out link with the right options' do
           expect(result).to eq(
             [
-              { text: 'Sign out', href: '/facilities-management/admin/sign-out', attributes: { method: :delete } }
+              { text: 'Sign out', href: '/facilities-management/admin/sign-out', method: :delete }
             ]
           )
         end
@@ -91,7 +91,7 @@ RSpec.describe HeaderNavigationLinksHelper do
         it 'returns the sign out link with the right options' do
           expect(result).to eq(
             [
-              { text: 'Sign out', href: '/facilities-management/supplier/sign-out', attributes: { method: :delete } }
+              { text: 'Sign out', href: '/facilities-management/supplier/sign-out', method: :delete }
             ]
           )
         end
@@ -104,7 +104,7 @@ RSpec.describe HeaderNavigationLinksHelper do
         it 'returns the sign out link with the right options' do
           expect(result).to eq(
             [
-              { text: 'Sign out', href: '/crown-marketplace/sign-out', attributes: { method: :delete } }
+              { text: 'Sign out', href: '/crown-marketplace/sign-out', method: :delete }
             ]
           )
         end